#881: Changed ResultsSource into a simple type and removed code for adding and removing event listeners. Now the ScriptureResultsViewer component just takes the array of sources and updates when the array changes.
The CheckingResultsList (in core) uses that component similar to before but now has logic to force a re-render when rerun is called, since it actually modifies the data in the sources rather than the array of sources itself.
However, this component is no longer consistently rendering the Rerun button. Seems there may be a timing issue as to whether the onRerun method is set at the time the view displaying the component is rendered.