Removed duplicate localization IDs and fixed a couple Spanish strings
For IDs defined in both assests\en.json and extensions\src\platform-get-resources\contributions\localizedStrings.json, I ensured they are only in localizedStrings.json if they are used only in that extension. Any IDs used elsewhere in Platform core/test code are defined in en.json
In the case of %resources_deprecated_any%, I decided to put it in en.json since it is not (and should not be) used in the extension.