PT-3394: Localized Platform UI strings into Spanish. (#1818)
* Localized most of the UI into Spanish.

Fixed a few small I18n issues.
Tweaked a couple English strings in minor ways.
Sorted the localization keys in the various files I touched.

* Translated new string in es.json and all the strings in new platform-lexical-tools extension

* PT-3394: Fixed GET_RESOURCES_STRING_KEYS in get-resources.web-view.tsx to include two new string IDs

* Removed duplicate localization IDs and fixed a couple Spanish strings

For IDs defined in both assests\en.json and extensions\src\platform-get-resources\contributions\localizedStrings.json, I ensured they are only in localizedStrings.json if they are used only in that extension. Any IDs used elsewhere in Platform core/test code are defined in en.json
In the case of %resources_deprecated_any%, I decided to put it in en.json since it is not (and should not be) used in the extension.

* Added Spanish localizations for new strings

Sorted out-of-order English localization IDs.